// src/types/rank.ts
export interface IRankItem {
  subscribers: any;
  subscribed: any;
  creator: string;
  artists: string;
  tracks: Array<{ first: string; second: string }>;
  updateFrequency: string;
  backgroundCoverId: number;
  backgroundCoverUrl: string;
  titleImage: number;
  titleImageUrl: string;
  englishTitle: string;
  opRecommend: boolean;
  recommendInfo: string;
  socialPlaylistCover: any;
  adType: number;
  trackNumberUpdateTime: number;
  subscribedCount: number;
  cloudTrackCount: number;
  userId: number;
  highQuality: boolean;
  createTime: number;
  specialType: number;
  coverImgId: number;
  updateTime: number;
  anonimous: boolean;
  newImported: boolean;
  coverImgUrl: string;
  trackUpdateTime: number;
  trackCount: number;
  totalDuration: number;
  commentThreadId: string;
  privacy: number;
  playCount: number;
  ordered: boolean;
  tags: string[];
  description: string;
  status: number;
  name: string;
  id: number;
  coverImgId_str: string;
  ToplistType?: string;
}